requirements.txt
discord.py==1.7.3
Pillow==9.0.0
python-dotenv==0.19.2
requests==2.26.0
